# src/tests/EshopGuard.Api.Tests/Snapshots/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ConsumeLoginLinkRequest": {
        "properties": {
          "market": {
            "type": [
              "null",
              "string"
            ]
          },
          "token": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "token",
          "market"
        ],
        "type": "object"
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "market": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "email",
          "market"
        ],
        "type": "object"
      },
...
      "LoginLinkInfoDto": {
        "properties": {
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "isNewAccount": {
            "type": "boolean"
          }
        },
        "required": [
          "email",
          "isNewAccount",
          "expiresAt"
        ],
        "type": "object"
      },
      "LoginLinkRequest": {
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "market": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "email",
          "market"
        ],
        "type": "object"
      },
      "LoginLinkRequestedDto": {
        "properties": {
          "expiresInSeconds": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          },
          "resendAfterSeconds": {
            "format": "int32",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "required": [
          "expiresInSeconds",
          "resendAfterSeconds"
        ],
        "type": "object"
      },
...
      "MeDto": {
        "properties": {
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "effectiveLocale": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "googleLinked": {
            "type": "boolean"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "locale": {
            "type": [
              "null",
              "string"
            ]
          },
          "localeSource": {
            "type": "string"
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/MeMembershipDto"
            },
            "type": "array"
          },
          "pendingInvitations": {
            "items": {
              "$ref": "#/components/schemas/PendingInvitationDto"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "locale",
          "effectiveLocale",
          "localeSource",
          "hasPassword",
          "googleLinked",
          "memberships",
          "pendingInvitations"
        ],
        "type": "object"
      },
      "MeMembershipDto": {
        "properties": {
          "legalName": {
            "type": [
              "null",
              "string"
            ]
          },
          "role": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "tenantId": {
            "format": "uuid",
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          }
        },
        "required": [
          "tenantId",
          "tenantName",
          "legalName",
          "role",
          "status"
        ],
        "type": "object"
      },
...
      "PasswordLoginRequest": {
        "properties": {
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "PendingInvitationDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "invitationId": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          }
        },
        "required": [
          "invitationId",
          "tenantName",
          "role",
          "expiresAt"
        ],
        "type": "object"
      },
...
      "ResetLinkInfoDto": {
        "properties": {
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "expiresAt"
        ],
        "type": "object"
      },
      "ResetPasswordRequest": {
        "properties": {
          "newPassword": {
            "type": [
              "null",
              "string"
            ]
          },
          "token": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "token",
          "newPassword"
        ],
        "type": "object"
      },
      "SessionDto": {
        "properties": {
          "createdTenantId": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "isNewAccount": {
            "type": "boolean"
          },
          "me": {
            "$ref": "#/components/schemas/MeDto"
          }
        },
        "required": [
          "me",
          "isNewAccount",
          "createdTenantId"
        ],
        "type": "object"
      },
...
      "TokenRequest": {
        "properties": {
          "token": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "required": [
          "token"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EshopGuard.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/auth/google/start": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "returnPath",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "market",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "csrf.invalid",
          "internal_error",
          "not_found",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "return_path.invalid"
        ]
      }
    },
    "/api/auth/login-link": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/LoginLinkRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginLinkRequestedDto"
                }
              }
            },
            "description": "Accepted"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "csrf.invalid",
          "email.send_failed",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
    "/api/auth/login-link/consume": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ConsumeLoginLinkRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "auth.account_disabled",
          "csrf.invalid",
          "internal_error",
          "login_link.expired",
          "login_link.invalid",
          "login_link.used",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
    "/api/auth/login-link/inspect": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/TokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginLinkInfoDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "csrf.invalid",
          "internal_error",
          "login_link.expired",
          "login_link.invalid",
          "login_link.used",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
...
    "/api/auth/password/forgot": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ForgotPasswordRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "csrf.invalid",
          "email.send_failed",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
    "/api/auth/password/login": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/PasswordLoginRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "auth.invalid_credentials",
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "validation.failed"
        ]
      }
    },
    "/api/auth/password/reset": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ResetPasswordRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "reset_link.expired",
          "reset_link.invalid",
          "reset_link.used",
          "validation.failed"
        ]
      }
    },
    "/api/auth/password/reset/inspect": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/TokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetLinkInfoDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "auth"
        ],
        "x-problem-codes": [
          "csrf.invalid",
          "internal_error",
          "rate_limited",
          "request.https_required",
          "request.invalid",
          "reset_link.expired",
          "reset_link.invalid",
          "reset_link.used",
          "validation.failed"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
...
}